hardware/pyproject.toml
[project]
name = "mpc-hardware"
version = "0.1.0"
description = "Crazyflie hardware control scripts"
requires-python = ">=3.13"
dependencies = [
    "cflib",
    "mpc-landing",
    "paho-mqtt",
    "pynput",
    "pyvista",
    "scipy",
]

[tool.setuptools]
py-modules = ["keyboard_control", "thrust_test", "mqtt_viewer", "mpc_control", "hover_test"]

[project.scripts]
keyboard-control = "keyboard_control:main"
thrust-test = "thrust_test:main"
mqtt-viewer = "mqtt_viewer:main"
mpc-control = "mpc_control:main"
hover-test = "hover_test:main"

[tool.uv.sources]
mpc-landing = { workspace = true }
